requirements.txt
# Core libraries for SQLi Hunter
rich
playwright
beautifulsoup4
simhash
dnspython
httpx
sqlglot
scikit-optimize
numpy # scikit-optimize requires numpy
playwright-stealth
curl_cffi
cloudscraper
PyQt6
PyQt6-sip
pyahocorasick
PyYAML

# Added for advanced features
torch
torchvision
torchaudio
torch_geometric
transformers
stable-baselines3[extra]
pyzmq
ray
qiskit
scikit-learn
h2
aioquic
websockets
